conver to more functional style
With the same functionality make more functional and less OO which
simplies the code despite being functionally the same

```
 src/catrr/__init__.py    | 160 ++++++++++++++++++++++---------------------------------------------
 src/catrr/cli.py         |  13 +++---
 tests/test_cli.py        |   2 +
 tests/test_rr.py         |  31 +++----------
 tests/test_statefulrr.py |  49 ---------------------
 tests/test_storage.py    |  52 ++++++++--------------
 6 files changed, 84 insertions(+), 223 deletions(-)
```

While doing this I realized that the actual items don't need the be
stored in the state file. The caller already has the items and the state
file need only worry about the hash and the current index. So this
commit also significantly reduces the amount of storage in the state
file.